+++ src/parser/util.ts
@@ -1,0 +1,3 @@
+export function isUnique<T>(value: T, index: number, self: T[]): boolean {
+  return self.indexOf(value) === index;
+}