+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Handle exports from within a namespace. Disabled by default, unless option --searchNamespaces is given. Note: this can affect performance on large codebases.
+
 ### Changed
 - (Internal) Update dependency TypeScript to 3.7.3
 - (Internal) Simplify some logic, using the new optional chaining operator (?.)